Temporary fix to the MovementApi spec
The Elite2 API endpoint which returns an offender's movement has been
recently updated, but only now causing an issue with our specs.  We have
a Jira ticket to update our codebase to work with the fact that we'll be
able to get all movements for an offender, but for now we need to
temporarily fix our specs to get other pieces of work deployed.