Fix Zendesk tag issue
Whilst testing the new Zendesk feature on staging we found an issue
where the tag for 'moic' wasn't always being set, which caused a problem
when trying to find relevant tickets.  In addition to this, the custom
fields weren't set up correctly and so the crucial information we were
collecting in the form wasn't then being translated to the Zendesk app.