Update MovementService
We have continued to see validation errors for 'prison can't be blank'
when running the movementService.  Another look at this showed that the
recently added method 'prison_fix' was in fact returning nil as the
movement_type check was against the string
'offender_released/transferred' rather than the constant which is being
set.