Currently, before we remove a co-working POM, we ask the user whether or not they want to allocate a new co-working POM by the use of YES/NO radio buttons.
In this commit we are removing the functionality to add a new
co-working POM when deleting the current co-working POM.
Instead we remove the co-worker and navigate back to the
allocation#show page where the user can add a new co-working POM
should they choose to.

The content for removing an allocated co-worker has also changed.