When fixing data, deallocate if we can't find an offender
Sometimes, when we attempt to get an offender from an allocation's
nomis_offender_id it does not exist. This typically means it has been
merged from an existing id for that offender.  Currently we can't look
this up, so now we will de-allocate the incorrect ID on the basis that
they will still appear for allocation under their new ID.