Fix POM working pattern bug
We have been receiving reports that the working pattern for POMs hasn't
been working properly.  Even if a POM was being added as working full
time the page would display 'Part time - 1.0', rather than 'Full time'.
It turns out that we're storing the working pattern as a float, but in
the helper we're comparing it to a string, and therefore it wasn't
working properly.  This PR updates the helper to compare against a float
and will now correctly display the working pattern.