Fix debugging page
A recent change to the MovementApi mean that we started getting all
movements for an offender, rather than just the latest.  The debugging
page displays the last movement, but as they're not returned in date
order it meant it could be showing an old movement.  Therefore this PR
adds the api call to get just the latest movement for an offender to
ensure the debugging page shows the right information.