Refactor MovementsApi
Following the review on the previous commit we realised that we had been
sorting all movements on the MovementTime field.  The issue being that
this is only returning a time, and so the ordering was broken as it
didn't factor in the dates.

This commit therefore switches to sort on the createDateTime field as it
contains the date and time.  It also means that we don't have near
duplicate code as can make use of the 'movements_for' method again.