Add allocation history link to Prisoner page
The prisoner information page that a POM accesses via their caseload
page needs to display a link to the allocation history page to reflect
the current service design.

As part of this PR I have also removed the last_event method from the
AllocationVersion model and added it to the OffenderHelper module as
really is a view helper.

I have also updated the "Back" link on the allocation history as it can
be accessed from two pages and it's important that the back link directs
the user back to the appropriate page.