Returns users back to the current paginated page after they have allocated or re-allocated an offender from the summary screen.
Changes have been made to the overrides_controller class.
Currently when a user performs an override the overrides_controller
passes back to allocations#confirm to update the allocations record
with the new POM. The event is saved as `allocate_primary_pom`.

The issue is that if the user is re-allocating a pom with override
reasons that record will be saved as `allocate_primary_pom` instead
of `reallocate_primary_pom`. For the purposes of this commit this
has the effect of returning the user to the "Make allocations" tab
instead of "See allocations" tab, which could be confusing to
users.

With this commit we are now identifying whether the event is an
allocation or reallocation within overrides_controller, returning to
either allocations#confirm or allocations#confirm_reallocation
respectively, allowing the flow of the journey to continue without
confusion.