Make handover date the responsibility of the sentence_detail
Currently the offender model is responsible for calculating the handover
date using local methods and the contents of the SentenceDetail object
stored in @sentence.

In cases where we don't have a whole offender (but do have the case
information for that offender) such as the caseload_controller it would
be better if the handover calculation could be done with just the case
information and sentence details.

This PR moves the responsibility for asking for handover dates to
SentenceDetail where it should be provided an instance of something that
responds to `nps_case?, early_allocation? mappa_level, and
indeterminate_sentence?).  The newly moved methods will use this object
along with the local dates to request the handover dates.