POM-363 Fix 'prison can't be blank' issue
We have seen a couple of new instances of the 'prison can't be blank'
error, which happens during the MovementService run.  Following an
investigation into this we realised that we had updated the MovementApi
method 'movements_for' to sort on the createDateTime field and therefore
need to check the last movement returned, not the first.  This PR
implements this fix and should resolve the issue.