POM-377 - Immigration Cases
Following a number of prisons contacting us regarding offenders who are
being held for immigration reasons and NOT showing up in the service for
allocation, Laura spoke with those involved with the OMIC policy for
guidance.  We have now had it confirmed that all offenders held for
immigration reasons should show up in the service and be allocated to a
POM.  These cases have the following rules:
- The POM will always have a supporting role
- The recommended POM will be a Prison POM.

This PR implements these rules and ensures that these cases are no
longer filtered out so they will start appearing in MOIC for allocation.