Change open prison responsibility and notify LDU
NPS Offenders who are in open prisons should be Supporting/Community. As
the offender moves into the open prison, their LDU should be notified
that they have to take responsibility for that offender - should the
offender have an LDU with a valid email address.  This is done during
the processing of transfers to check if the target prison is open or
not. If it is, then an email is sent to the LDU with some offender
details.

As a drive-by, this PR also ups the max CyclomaticComplexity to 10 so we
don't get punished just for using if statements.