Fix error where it was possible to edit PRD too early
Currently the link to edit the offender's PRD is available on the
case_information#show page - but unfortunately there is no case
information to edit at that point.

This PR adds a local to wherever the partial is rendered specifying
whether it should be possible to edit the PRD from that page (and if so,
showing the link).

Parole review date was also duplicated in case information and so has
been removed and left in offence_info.