Ensure offenders returned have sentences and caseinfo
Previously this just returned the bare offender record, and now it
actually does the usual check to get the sentences and case info.

This chunk of code that fetches case info and sentences is now done in a
couple of places, so it may be a good candidate for refactoring how we
do this.