POM-432 Multiple POM records
We had been contacted by a number of SPOs advising that they were unable
to update their POM working patterns.  An investigation led to finding
17 POMs who have two PomDetail records, which meant that one record was
being updated, but the other being displayed with the incorrect data.
We are adding a uniqueness constraint to the nomis staff ID field in
both the db and model to prevent this happening again.