Fix single offender caching
Currently the code caches the Offender object, but this means if the
code underneath that changes when it is unpickled from the cache then
data may be missing, e.g. if a field name in the model code changes.

This PR flips to storing the JSON for the offender and de-serialising
after fetching it which means that it should pick up any model changes
if the deserialisation is sound.

The key is changed so that we don't have to manually clear the cache.