Stop typhoeus logging useless info
Currently we get a log line from typhoeus in addition to a line logging
the http(s) calls to Elite2.

This PR removes the lines starting ETHON but leaves us with lines that
look like...

```
[gateway.t3.nomis-api.hmpps.dsd.io] GET
/elite2api/api/locations/description/LEI/inmates?convictedStatus=Convicted&returnCategory=true
(0.554 s)
```

Which is all we really need.  We can fix this up if we ever get around
to structured logging.