OpenPrisonTransferJob fails with nil:NilClass error
The error occurs within the following code
	offender = OffenderService.get_offender(movement.offender_no)
	return unless offender.nps_case?

We assumed that the parameter movement_json in OpenPrisonTransferJob.perform method
was in camel case format. So we call Nomis::Movement.from_json to convert it to
snake case format. But movement_json was already in snake case format so calling
Nomis::Movement.from_json, results in nil values for all of the keys (because the
from_json method is extracting values from camel case format hash keys.
These keys do not exist in movement_json as the keys are in snake case format)

So currently in the code above we were actually calling
	offender = OffenderService.get_offender(nil)

Anything after this point will end up resulting in the NilClass error.

This commit fixes the error by calling Nomis::Movement.new. Because movement_json
is already in the right format we can just use Nomis::Movement.new to convert
the json into Movement instance variables, which means we can call
movement.offender_no and it will return an actual value instead of nil.

In this commit we are also handling the situation where calling
OffenderService.get_offender returns nil.