Provide faceting and filtering of a POM's caseload
An upcoming ticket focused on redesign of various screens has a
component that describes filters for a POM's caseload.  As part of that
component it shows a facet of the total number of allocations (post
application of filters if any).

This PR implements (currently) just the calculation of the facet counts
for each PomCaseload.