Show correct responsibility when it has been overridden
When an offender has their responsibility overridden to be PROBATION,
the currently allocated POM should show as SUPPORTING, and not
RESPONSIBLE.  Allocated secondary POM will become CO-WORKING.

Currently this was not handled by the POMCaseload or the previous
PrisonOffenderManagerService function.

This PR now checks for an overriddent responsibility when determining a
POMs caseload.