[SPIKE] Investigate how we can more easily generate metrics
For discussion

Currently we use the prometheus exporter gem which means we have another
process in another container collecting metrics and making them
available at a /metrics endpoint.  This however is not fun to work with.

This PR change how we collect metrics to use the Prometheus provided
Ruby client, which is a lot simpler and exposes a /metrics endpoint
within our application.

The MetricService (I know, I know) shows how we can create counters and
gauges and the SearchController shows how we would record them.

Metrics from sidekiq is a more complicate issue and so would need to
provide its own /metrics endpoint.  Trials with prometheus_sidekiq (from
fastly) were problematic because of a version conflict with the client.