Access to various controllers was not correctly enforced
POMs should not see SPO task links on the dashboard, and we should
ensure the user has the correct role to be able to access those
controllers.

Also changes search so that unless the user is an SPO, they will be
redirected to caseload search instead of the actionable search (which
allows allocating etc).