Do not let invalid allocations appear in a POM's caseload.
Currently the caseload for a POM gets all of their allocations and the
matching offenders. This currently includes invalid allocations, where
the offender is no longer convicted, or no longer at the same prison as
the POM/Allocation.

To avoid this we are now excluding allocated offenders from the caseload
where:

  * The offender is reported at a different prison/out
  * The offender is not convicted

We _do not_ de-allocate the offender (although we _could_ as long it was
constrained to any allocations at active prison), but we do log the
error with a [ALLOC] tag so that it is reasonably easy to locate them in
logs.

Currently this does not solve the problem, as the allocation still
exists (although if they are allocated at another prison then the
incorrect allocation will be corrected), so we still need to decide what
to do about those incorrect allocations as they may result in incorrect
numbers.

DISCUSS:  When we find an invalid allocation, we should do something
like:

```
DeallocateOffenderAtPrisonJob(prison, offender).perform_later
```