Contact us errors conforms to GDS style guidelines.
We no longer uses the shared/validation_errors

It was not possible to test the errors due to unpermitted params,
fixed this by adding the additional parameters when in test and
development.

Disabled turbolinks when linking to anchors as this stops the
page from being refreshed and losing the error messages.