OffenderPresenter has a missing function that should be delegated
The OffenderPresenter needs to delegate nps_case? for situations where
it is used for determining the POM recommendation.

This was not caught because the tests used a struct where this was
defined.  Adding this function should make the test reflect the object
and should enable open prisons to allocate offenders.