Update Redis & Sidekiq config
Whilst setting up the integration with Heroku it was discovered that it
is expecting the REDIS_URL to contain all the data it needs, including
the password.  However, due to our initial setup for our elasticache
instance on Cloud Platform our configuration needed the host and auth
split out. This was causing issues when trying to get Heroku set up as
it kept returning 401s when running the app.  It turns out that we can
now get the full redis URL from CP and therefore this PR updates the
code to make use of this, hopefully ensuring that both Heroku and CP
deployments can work with the same setup.