Show 'new arrivals' in summary
As per POM-450 this PR adds a new summary type (:new_arrivals) to the
summary, which is "offenders that would be in missing information if
they have been waiting more than 2 days".  To achieve this means we need
to include the arrival date for all offenders in all summary_types so
that we can calculate totals. This is not ideal.