[POM-518] Heroku - get Notify working
At present the sidekiq worker doesn't start once a branch/PR is deployed
and this means that email jobs get sent to the Sidekiq queue, but aren't
processed.

This commit tries another configuration which specifies the number of
workers we want in the app.json file.  It also removes the default
setting of workers from the Procfile, to try and avoid
conflict/confusion.