[POM-520] Stop crashing when POM not properly setup
https://dsdmoj.atlassian.net/browse/POM-520

At present we crash if a POM is not properly set up i.e. a POM needs to
have the ALLOC_CASE_MGR role, and have a POM job type in classic NOMIS.
We therefore need to handle this more effectively, and redirect an
incorrectly set up POM to a 401 page.