Next steps:
This commit updates the rules regarding responsibility calculation for
CRC cases, and brings in the new home_detention_curfew_actual_date date
from the sentence information we get back from NOMIS to help with this
calculation.

Specs have been updated to reflect the changes made in regards to the
dates that are being checked, and Rubocop has been appeased.

 50-character subject line