Finalise POM responsibility calculation
This PR introduces a check for determinate cases that do not have any of
the release dates we are now basing our calculations on (CRD, ARD,
HDCED, & PED), if all are returned as nil then the case will default to
being POM responsible.  Specs have been updated to reflect this change.