Fix responsibility override on debugging page
The 'responsibility overridden?' section of the debugging page hadn't
been updated to reflect the changes we've been making, and has been
displaying the answer to whether the recommended POM type was overridden
at allocation.  Therefore this PR updates the page to display whether
the case responsibility has been overridden to the community or not.