Add specs for 'new arrivals' w/longform mocks
This commit pulls out the 'new arrivals' cases and cuts them off
from the heavy mocks used in the rest of the spec.

There's some classes that mock out things in the application already,
but as a naive reader they seem to be causing unusual behaviour, so
these specs were built up one webmock at a time to help them mock the
minimal set of data needed for this feature.

This commit introduces an empty (and ignored) spec for movements,
which may or may not be part of the previously built feature on this
branch. This can be worked on in a different commit.

Notably, there is a lot of duplication in these specs that needs
addressing.