Fix pom-caseload-feature-spec
In the last couple of days one of the tests in the pom-caseload-feature
spec file has started to fail; related to sorting by release date and
expecting certain names of offenders to appear in a specific order.
This test started failing across all branches, including master and
following an investigation it appears that there may have been a change
in data in T3, which has caused this to start failing.  The longer term
plan is to start refactoring specs to use webmock, but for now we need
to get this test passing as it is blocking other PRs.