Pom 574 - Fix POM responsibility calculation (#1029)
* POM-574

We have been incorrectly calculating POM responsibility as needed to
take into account the length of time left to serve at the start of the
offender's sentence, and ensure that responsibility correctly flips to
supporting at the handover date.

* Appease the linter

* Update responsibility calc for:

NPS / determinate / English / New case - now correctly uses ARD/CRD
NPS / determinate / English / New case / Parole Eligibility - now
correctly uses PED

* Appease linter

* Next steps:

This commit updates the rules regarding responsibility calculation for
CRC cases, and brings in the new home_detention_curfew_actual_date date
from the sentence information we get back from NOMIS to help with this
calculation.

Specs have been updated to reflect the changes made in regards to the
dates that are being checked, and Rubocop has been appeased.

 50-character subject line

* Interim work

* Appease the linter

* Responsibility calc for NPS indeterminate cases

* Beginning stages of refactoring the spec tests for the ResponsibilityService.

We also have commented code to explain why we aren't calculating certain
cases.

* Pre-policy cases now use correct range of release dates (ARD, CRD or PED) and specs have been fleshed out

* Amended Handover calculation dates, still a work in progress

* Finalise POM responsibility calculation

This PR introduces a check for determinate cases that do not have any of
the release dates we are now basing our calculations on (CRD, ARD,
HDCED, & PED), if all are returned as nil then the case will default to
being POM responsible.  Specs have been updated to reflect this change.

* Update allocation service spec to use offender with relevant release dates

* Update comment in spec & remove no longer needed variable crd_ard

* Add cache invalidation to tests
The cache is cleared after
each test run. This is done because the response
of each request is stored in a cache and inorder to make another request with
the same key, the cache must be cleared. Aleternatively, you could wait by adding a sleep for the duration of time it takes for the cache to expire to make the request. This is at the expense of the speed of your test runs.

* Add acceptance test for custody responsibilty
This acceptance test only tests two unique cases. Not all edge cases or scenarios are tested.
There is a test which shows a POM responsible and a POM as supporting.

Co-authored-by: Andrien <andrien.ricketts@digital.justice.gov.uk>
Co-authored-by: Maysa Kanoni <46002877+maysakanoni@users.noreply.github.com>