Fix sorting by arrival date
The sorting by arrival date on the new arrivals page/tab isn't working
as expected.  Our belief is that the default sorting option included a
field that isn't present on the page and therefore could have been
causing the issue.  We've set a new set of fields to sort by.