Pom 450 new arrivals (#976)
* Show 'new arrivals' in summary

As per POM-450 this PR adds a new summary type (:new_arrivals) to the
summary, which is "offenders that would be in missing information if
they have been waiting more than 2 days".  To achieve this means we need
to include the arrival date for all offenders in all summary_types so
that we can calculate totals. This is not ideal.

* Remove unnecessary test

* Clarify the time period for newly arrived

* put back breadcrumbs

* fix to only return new arrivals if they are <= 2 days, as by the time 2 days has elapsed (e.g. they arrive on Monday, and its now Wednesday) nDelius matching should have happened.

* improve test slightly

* improve test, added comment about tier presence implying case information presence

* improved tests

* tidied up Summary class - remove writers

* put back half of the feedback feature test

* tidy up template

* fixed bug with offenders arriving on Fridays

* fixed bug where saturday offender not counted as new arrival on Monday

* appease the linter

* Finish resolving issues related to rebasing master

* Implement 'travel_to' helper method in summary_controller_spec to try and get tests passing again

* WIP better time travel in specs

* Add specs for 'new arrivals' w/longform mocks

This commit pulls out the 'new arrivals' cases and cuts them off
from the heavy mocks used in the rest of the spec.

There's some classes that mock out things in the application already,
but as a naive reader they seem to be causing unusual behaviour, so
these specs were built up one webmock at a time to help them mock the
minimal set of data needed for this feature.

This commit introduces an empty (and ignored) spec for movements,
which may or may not be part of the previously built feature on this
branch. This can be worked on in a different commit.

Notably, there is a lot of duplication in these specs that needs
addressing.

* Bust cache between spec runs

* Add spec to cover scenario where previously allocated offender is transferred.  Other small changes in line with Rubocop

* Undo deletion of the beta phase banner

* Add newline to end of file

* temporary preprod change

* Fix sorting by arrival date

The sorting by arrival date on the new arrivals page/tab isn't working
as expected.  Our belief is that the default sorting option included a
field that isn't present on the page and therefore could have been
causing the issue.  We've set a new set of fields to sort by.

* temporarily disable rubocop Metrics/PerceivedComplexity in summary service

* Update new arrivals sorting fields to include correctly named 'prison_arrival_date'

* Revert temporary changes made for testing to take place; reinstate tests and reset the arrival date grace period

* Readd phase banner to errors amd contact views

* update test to reflect change in service
This test is calling the api. Although this isnt the best approach and the reason for an
offender disappearing is unclear- the quick and easy approach is to update the
test to reflect this change. Eventually, tests will not be calling a live
service.

Co-authored-by: Stephen Dicks <github@starswan.com>
Co-authored-by: kath <kath.pobee-norris@digital.justice.gov.uk>
Co-authored-by: Kyle <kjdchapman@gmail.com>
Co-authored-by: Maysa Kanoni <46002877+maysakanoni@users.noreply.github.com>