Update Handover calculation to catch more cases
Following the deployment of the responsibility and handover calculations
we started getting errors in regards to handover dates failing due to
determinate offenders not having conditional release dates, automatic
release dates or parole eligibilty dates.  This is a temporary fix to
catch these cases, and display no handover date due to not having
'earliest release date'. We will then need to decide how to proceed with
these cases moving forwards.