Take leap years into account for offender age
Whilst writing tests to cover this method, as we're hoping to change
it to handle `nil` a bit better, we realised the implementation does
not take into account leap years, so returns the wrong age if an
offenders' birthday is particularly close.

The new implementation is a little longer, but more accurate (and
hopefully more expressive).