Add cache invalidation to tests
The cache is cleared after each test run. This is done because the
response of each request is stored in a cache, which can cause
confusing behaviour for other tests ('leaky' tests).

Previously, some tests use sleeps to get around the cache - this
is at the expense of the speed of your test runs.

Co-authored-by: Maysa <maysa@madetech.com>