Show correct next delius import day
Similarly to 9deba6, this commit changes a summary helper
(responsible for displaying information about new arrivals) to test
equality on the base of calendar days, not exact time-stamp.

This commit also refactors the related tests, and adds new ones to
cover the case of more specific timestamps being in `arrival_date`.