Process movements for offenders moving to immigration centres
This PR is the result of a bug that was showing incorrect numbers of Prison Offender Manager (POM)
allocations in a particular prison. Looking into this issue we were able to identify that a couple
of POM's had not been de-allocated when an offender had been released or transferred,
and this was resulting in the incorrect POM allocation counts.

The MovementService handles de-allocating POMs when an offender moves prison, so we ran a check across
all prisons to find out if there were more situations where POM's were showing up as allocated to an
offender that was no longer at the POM's prison.

We found approximately 140 other cases and the vast majority were due to offenders moving from prison
to an immigration centre, awaiting deportation. The MovementService did not process these movements as
immigration centres are not part of the prison estate and as a result these offenders were not being
de-allocated from their POM when they moved.

This commit intends to fix this issue with the MovementService by de-allocating the POM and deleting the
CaseInformation when the offender is being transferred or released to an immigration centre - signified
by the agency code "IMM" (Immigration) or "MHI" (Morton Hall Immigration Removal Centre).