Use existing configuration for paths in spec
The previous approach - setting the Rails config to match some
fictional path - worked for this spec, but broke other specs in the
test suite, as the configuration change leaks between tests.

It's possible to set the configuration then reverse the change in
before/after blocks, but this adds so many lines to the test that I
think it's not worth the tradeoff.

I *think* this implementation will break if these config flags are
missing entirely, but this'll break the entire test suite in a pretty
obvious way, so we're not making the problem worse with this spec,
at least.