Prevent errors on missing dates in handover calcs
This commit changes the handover service to return `nil` instead of
raising a error when essential dates are missing. It also adds specs
around this behaviour.

Notably, some of this behaviour is not covered by acceptance specs,
but is also not yet used in any user-facing parts of the service.

It also updates an acceptance spec around the allocations page, which
was where the original crash was identified, to match the data which
caused that original crash.