Return no age for an offender if dob is nil
Whilst this case should "never happen", we want sensible output if it
does. Notably this is _quieter_ than the previous error, so needs
followup work to display that certain information is missing to users.

This commit also changes a filter, based on age, to ignore offenders
who do not have a dob.