Return no staff name if first/last names are nil
Similarly to da16d1, this data "should" never be missing, but it's
better to display it as missing rather than throwing an error.

Unfortunately, the related test (spec/models/staff_member_spec.rb)
proved to be difficult to extend to cover this case. Given the intent
is to continue writing acceptance specs that cover this behaviour,
it feels like the correct trade-off to avoid that problem for now.