Handle missing dates in responsibility service
This commit introduces some guard clauses to check if dates are nil
before comparing them to thresholds. If they're nil, then the service
will also return nil (translating to "we don't know who should be
responsible"). It also adds specs (both unit and acceptance) to cover
this new behaviour.

It also does a little bit of refactoring on the responsibility
service, to make the guard clause ("get out of here if these dates are
nil!) more obvious.