Handle incorrectly entered service provider
When case information records are created for offenders there is a
fallback option for prison staff to enter this data correctly.  Whilst
testing the responsibility and handover date services we uncovered a
case where an indeterminate offender had been incorrectly labeled as a
CRC case, rather than an NPS one.  This meant that the calculations
failed due to the case falling into the CRC path.  This PR adds an extra
check for indeterminate cases and pushes them down the NPS route even if
they have been set as a CRC case in error.