Fix recall calculations
We received a Zendesk ticket from a prison alerting us to a recalled
offender that was still showing as the POM being responsible.  An
investigation led us to realise that this particular offender didn't
have the relevant release dates and therefore was defaulting to
responsible.  We have re-ordered the calculation list to ensure that
recalled offenders drop out/default to supporting before other scenarios
are taken into account.