Fix issue when co-working POM no longer a POM
We recently received a report from an SPO where when trying to view an
offender's allocation that they were being redirected to the /401 page.
Following an investigation it was discovered that the co-working POM was
no longer an active POM at the prison, and it wasn't redirecting to the
page to notify the SPO/letting them deallocate their cases.  It appears
that a small typo had caused this issue so looking to resolve this issue
with this PR.