POM-615 - Set user expectations
Prison staff are able to contact the team by completing a form on the
help page, which then raises a Zendesk ticket.  This PR updates the
flash message shown to the user, which now advises them that they will
be contacted within 2 working days.  This is part of a larger piece of
work around setting, and managing, user expectations.