POM-637 Get POM name from NOMIS
We have received a report from a user advising that one of their POM’s
has changed their name, it’s been updated in NOMIS but it’s still
showing up as the old name on DPS.  A quick investigation unearthed that
when DPS calls the MOIC API we are sending back the POM name in the
allocation record, and not from NOMIS, so it won’t show the correct name
if the POM has changed their name.  We need to update this, and other
places where we display the POM name in MOIC to fetch the name from
NOMIS to ensure it’s accurate/up to date.