Improve clarity around release dates displayed
Since starting MOIC we have iterated in regards to the different dates
we use to determine if an offender is sentenced, calculating POM
responsibility and Handover dates.  However, we haven't updated the
dates we display to reflect these changes, and it has led to some
confusion for users.

This PR focusses on removing all references to 'release date' as other
than using this to determine if an offender is sentenced we do not use
it.  In addition, we have updated the dates we use when calculating
responsibility and handover, so the 'earliest release date' method now
matches these.

We are looking to replace the release date with the sentence start date
on some of the pages within the service.  There are times when offenders
sentences haven't been entered correctly on NOMIS and this can lead to
their sentence start date being wrong/absent.  The hope is that by
exposing this data to SPO/POMs that they can fix this data themselves
rather than contacting us via Zendesk.

I have also been bad and snuck in some changes that Rubocop required as
it was complaining about an incorrectly named cop.