Improve Earliest Release Date
There are offenders where one or more of the dates we use to calculate
the 'earliest release date' falls into the past, and yet have another
date set in the future.  It has been agreed that a sensible approach to
earliest_release_date is to iterate over the dates we get back and
determine if they're in the future or the past.  If the offender has any
dates in the future we will choose the earliest of those, and if they
only have dates in the past we will then default to the earliest of
these.  This should avoid us displaying release related dates that have
already past, but where an offender also has other dates in the future.