DRY up inputs/result in responsibility specs
This commit pulls out the various 'offender' mocks, uses an rspec
double to draw attention to the role of this object (instead of an
OpenStruct), and de-dupes a bunch of stuff to reduce noise.

I'm not happy with l91-100, and I'm sure there's a better way to
structure it, but I believe it's a step up from having a double
repeatedly set up across many different context blocks.