Keep POMs responsible w/far-in-advance MAPPA flag
A case was spotted where an offender was due to be released some
time away (longer than the 7.5 months of the handover process), but
was displaying *today* as their handover date. This also meant their
responsibility was showing as community rather than custody.

This was due to a 'MAPPA' flag being set in nDelius, which our service
takes to mean "this person is being released imminently, so hand them
over now". In actuality, this flag can be set far in advance of a
handover process (even if it "shouldn't" be set), so this can lead
our service to display confusing/incorrect information.

Our domain expert suspects this is something that will happen a fair
amount, so changing our service to respond sensibly is a good idea.

This commit only fixes the MAPPA case explicitly, but it would be
good to enshrine the domain rule of *the date we're changing
responsibility from custody to community can never be before the date
we're planning to start handover* in code, in a later commit. This
will probably require further refactoring of the
`handover_date_service`, however.