Handle missing release dates in handover calc
This case of missing data was missed in the changes in 39134a - if
`max` is called on an array with nil and non-nil entries, it'll
explode.

This commit `compacts` the array first as a quick fix - the idea of
'if there is missing data behave this way' could do with being pulled
up into clearer, separate components in the future, as this keeps
biting me.