Update process of POM being made inactive
At present an SPO is able to make a POM inactive by updating the POM's
profile, and once done this will deallocate the POM's cases.  However,
at present this is only deallocating the cases where they are the
primary POM, resulting in allocations which continue to have an inactive
POM associated to them.  Therefore this PR updates the process and
ensures that the POM is removed all allocations, both as a primary and
secondary POM.